Add frame around windows for management
Operations to move/resize windows are done directly from the title bar,
at the top of the window. No modmask is needed when clicking them.

Because it works by reparenting the windows, the focus needs to be set
explicitely to the first child of each frame when the mouse enters them.

Same goes for destroying, which occurs when the child window gets deleted.