Make cb_create more robust
When a new window is created, we won't decorate it if it is a frame
window, or if the overrid_redirect bit is set.
We also need to move it somewhere near the cursor to avoid them popping
at the top left, with the decoration outside the screen (which makes it
impossible to drag).