Add new callback for CONFIGURE events
The goal here is to resize the frame window whenever it's direct child
is resized/moved.
When such an event occur, the frame will receive a "CONFIGURE_NOTIFY"
event with the new dimension/position of the window.

We need to take extra care with such events, because when we resize
the frame window, we also force resizing the child, and that triggers
a configure event. The quick and dirty way is to not handle these events
when a window is selected (`curwid` is set) and being dragged with
the cursor.