Handle mouse press events on every window
When a mouse button is pressed, the .event element of the structure holds
the ID of the window that received the event, not especially the window
that received the mouse click.

This commit will save the window that received the click in a variable
for later use, rather than assuming that the window receiving the event
and the window receiving the click are the same.